Refactor project into src layout and move tests
