+++ pyproject.toml
@@ -1,0 +1,10 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "assist"
+version = "0.1.0"
+
+[tool.setuptools.packages.find]
+where = ["src"]
