+++ src/assist/tools/project_index.py
